+++ ocfl-java-core/src/test/java/io/ocfl/core/inventory/InventoryMapperTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import io.ocfl.api.model.DigestAlgorithm;
+import io.ocfl.api.DigestAlgorithmRegistry;
 import io.ocfl.core.model.Inventory;
 import io.ocfl.core.model.RevisionNum;
 import io.ocfl.core.test.ITestHelper;
@@ -33,7 +33,8 @@ public class InventoryMapperTest {
         var objectRoot = "path/to/obj1";
         var digest =
                 "bd9b8dff3b9b3debe2f5f88023f85bb501603711861b7e3da093fc970e149a0972797b3e03080e43a75974b51420b269dc87cd7bd64836f676ef06844b2ef345";
-        var inventory = mapper.read(objectRoot, DigestAlgorithm.sha512, new ByteArrayInputStream(original.getBytes()));
+        var inventory =
+                mapper.read(objectRoot, DigestAlgorithmRegistry.sha512, new ByteArrayInputStream(original.getBytes()));
         assertFalse(inventory.hasMutableHead());
         assertNull(inventory.getRevisionNum());
         var output = writeInventoryToString(inventory);
@@ -50,7 +51,7 @@ public class InventoryMapperTest {
                 "bd9b8dff3b9b3debe2f5f88023f85bb501603711861b7e3da093fc970e149a0972797b3e03080e43a75974b51420b269dc87cd7bd64836f676ef06844b2ef345";
         var revision = RevisionNum.fromString("r2");
         var inventory = mapper.readMutableHead(
-                objectRoot, revision, DigestAlgorithm.sha512, new ByteArrayInputStream(original.getBytes()));
+                objectRoot, revision, DigestAlgorithmRegistry.sha512, new ByteArrayInputStream(original.getBytes()));
         assertTrue(inventory.hasMutableHead());
         assertEquals(revision, inventory.getRevisionNum());
         var output = writeInventoryToString(inventory);