+++ ocfl-java-itest/src/test/java/io/ocfl/itest/OcflITest.java
@@ -24,6 +24,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
+import io.ocfl.api.DigestAlgorithmRegistry;
 import io.ocfl.api.OcflConfig;
 import io.ocfl.api.OcflOption;
 import io.ocfl.api.OcflRepository;
@@ -149,20 +150,20 @@ public abstract class OcflITest {
         var file1Sha512 =
                 "6407d5ecc067dad1a2a3c75d088ecdab97d4df5a580a3bbc1b190ad988cea529b92eab11131fd2f5c0b40fa5891eec979e7e5e96b6bed38e6dddde7a20722345";
         var inputStream1 = new FixityCheckInputStream(
-                ITestHelper.streamString(file1Contents), DigestAlgorithm.sha512, file1Sha512);
+                ITestHelper.streamString(file1Contents), DigestAlgorithmRegistry.sha512, file1Sha512);
 
         var file2Contents = "... contents of second file ...";
         var file2Sha512 =
                 "d173736e7984e4439cab8d0bd6665e8f9a3aefc4d518a5ed5a3a46e05da40fa5803ac5dc52c9b17d302e12525619a9b6076f33a0c80b558bff051812800e0875";
         var inputStream2 = new FixityCheckInputStream(
-                ITestHelper.streamString(file2Contents), DigestAlgorithm.sha512, file2Sha512);
+                ITestHelper.streamString(file2Contents), DigestAlgorithmRegistry.sha512, file2Sha512);
         inputStream2.enableFixityCheck(false);
 
         var file3Contents = "... contents of third file ...";
         var file3Sha512 =
                 "f280e67f4142469ac514dd7ad366c6ed629e10b30f6f637e6de36b861c44ba5753d8fe8d589b9b23310df9e9d564a20a06d5f4637bd9f8e66ab628c7cce33e72";
         var inputStream3 = new DigestInputStream(
-                ITestHelper.streamString(file3Contents), DigestAlgorithm.sha512.getMessageDigest());
+                ITestHelper.streamString(file3Contents), DigestAlgorithmRegistry.sha512.getMessageDigest());
 
         repo.updateObject(ObjectVersionId.head(objectId), new VersionInfo(), updater -> {
             updater.writeFile(inputStream1, "file1");
@@ -175,17 +176,17 @@ public abstract class OcflITest {
         try (var stream = object.getFile("file1").getStream()) {
             assertEquals(file1Contents, new String(stream.readAllBytes()));
         }
-        assertEquals(file1Sha512, object.getFile("file1").getFixity().get(DigestAlgorithm.sha512));
+        assertEquals(file1Sha512, object.getFile("file1").getFixity().get(DigestAlgorithmRegistry.sha512));
 
         try (var stream = object.getFile("file2").getStream()) {
             assertEquals(file2Contents, new String(stream.readAllBytes()));
         }
-        assertEquals(file2Sha512, object.getFile("file2").getFixity().get(DigestAlgorithm.sha512));
+        assertEquals(file2Sha512, object.getFile("file2").getFixity().get(DigestAlgorithmRegistry.sha512));
 
         try (var stream = object.getFile("file3").getStream()) {
             assertEquals(file3Contents, new String(stream.readAllBytes()));
         }
-        assertEquals(file3Sha512, object.getFile("file3").getFixity().get(DigestAlgorithm.sha512));
+        assertEquals(file3Sha512, object.getFile("file3").getFixity().get(DigestAlgorithmRegistry.sha512));
     }
 
     @Test
@@ -391,28 +392,28 @@ public abstract class OcflITest {
 
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("1"), updater -> {
             updater.addPath(ITestHelper.sourceObjectPath(objectId, "v1"))
-                    .addFileFixity("file1", DigestAlgorithm.md5, "95efdf0764d92207b4698025f2518456")
-                    .addFileFixity("file2", DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad");
+                    .addFileFixity("file1", DigestAlgorithmRegistry.md5, "95efdf0764d92207b4698025f2518456")
+                    .addFileFixity("file2", DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad");
         });
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("2"), updater -> {
             updater.clearVersionState()
                     .addPath(ITestHelper.sourceObjectPath(objectId, "v2"))
-                    .addFileFixity("file1", DigestAlgorithm.md5, "a0a8bfbf51b81caf7aa5be00f5e26669")
-                    .addFileFixity("file2", DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad")
-                    .addFileFixity("dir1/file3", DigestAlgorithm.md5, "72b6193fe19ec99c692eba5c798e6bdf");
+                    .addFileFixity("file1", DigestAlgorithmRegistry.md5, "a0a8bfbf51b81caf7aa5be00f5e26669")
+                    .addFileFixity("file2", DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad")
+                    .addFileFixity("dir1/file3", DigestAlgorithmRegistry.md5, "72b6193fe19ec99c692eba5c798e6bdf");
         });
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("3"), updater -> {
             updater.clearVersionState()
                     .addPath(ITestHelper.sourceObjectPath(objectId, "v3"))
-                    .addFileFixity("file2", DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad")
-                    .addFileFixity("file4", DigestAlgorithm.md5, "a0a8bfbf51b81caf7aa5be00f5e26669");
+                    .addFileFixity("file2", DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad")
+                    .addFileFixity("file4", DigestAlgorithmRegistry.md5, "a0a8bfbf51b81caf7aa5be00f5e26669");
         });
 
         var objectDetails = repo.describeObject(objectId);
 
         assertEquals(objectId, objectDetails.getId());
         assertEquals(VersionNum.fromString("v3"), objectDetails.getHeadVersionNum());
-        assertEquals(DigestAlgorithm.sha512, objectDetails.getDigestAlgorithm());
+        assertEquals(DigestAlgorithmRegistry.sha512, objectDetails.getDigestAlgorithm());
         assertEquals(3, objectDetails.getVersionMap().size());
 
         assertThat(
@@ -425,16 +426,16 @@ public abstract class OcflITest {
                                 "file1",
                                 O1_PATH + "/v1/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "96a26e7629b55187f9ba3edc4acc940495d582093b8a88cb1f0303cf3399fe6b1f5283d76dfd561fc401a0cdf878c5aad9f2d6e7e2d9ceee678757bb5d95c39e",
-                                        DigestAlgorithm.md5, "95efdf0764d92207b4698025f2518456")),
+                                        DigestAlgorithmRegistry.md5, "95efdf0764d92207b4698025f2518456")),
                         fileDetails(
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6",
-                                        DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad"))));
+                                        DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad"))));
 
         assertThat(
                 objectDetails.getVersion(VersionNum.fromString("v2")),
@@ -446,23 +447,23 @@ public abstract class OcflITest {
                                 "file1",
                                 O1_PATH + "/v2/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc",
-                                        DigestAlgorithm.md5, "a0a8bfbf51b81caf7aa5be00f5e26669")),
+                                        DigestAlgorithmRegistry.md5, "a0a8bfbf51b81caf7aa5be00f5e26669")),
                         fileDetails(
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6",
-                                        DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad")),
+                                        DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad")),
                         fileDetails(
                                 "dir1/file3",
                                 O1_PATH + "/v2/content/dir1/file3",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "cb6f4f7b3d3eef05d3d0327335071d14c120e065fa43364690fea47d456e146dd334d78d35f73926067d0bf46f122ea026508954b71e8e25c351ff75c993c2b2",
-                                        DigestAlgorithm.md5, "72b6193fe19ec99c692eba5c798e6bdf"))));
+                                        DigestAlgorithmRegistry.md5, "72b6193fe19ec99c692eba5c798e6bdf"))));
 
         assertThat(
                 objectDetails.getVersion(VersionNum.fromString("v3")),
@@ -474,16 +475,16 @@ public abstract class OcflITest {
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6",
-                                        DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad")),
+                                        DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad")),
                         fileDetails(
                                 "file4",
                                 O1_PATH + "/v2/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                                 "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc",
-                                        DigestAlgorithm.md5, "a0a8bfbf51b81caf7aa5be00f5e26669"))));
+                                        DigestAlgorithmRegistry.md5, "a0a8bfbf51b81caf7aa5be00f5e26669"))));
 
         assertSame(objectDetails.getHeadVersion(), objectDetails.getVersion(VersionNum.fromString("v3")));
     }
@@ -499,11 +500,11 @@ public abstract class OcflITest {
             updater.addPath(ITestHelper.sourceObjectPath(objectId, "v1"))
                     .addFileFixity(
                             "file1",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "96a26e7629b55187f9ba3edc4acc940495d582093b8a88cb1f0303cf3399fe6b1f5283d76dfd561fc401a0cdf878c5aad9f2d6e7e2d9ceee678757bb5d95c39e")
                     .addFileFixity(
                             "file2",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6");
         });
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("2"), updater -> {
@@ -511,15 +512,15 @@ public abstract class OcflITest {
                     .addPath(ITestHelper.sourceObjectPath(objectId, "v2"))
                     .addFileFixity(
                             "file1",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc")
                     .addFileFixity(
                             "file2",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")
                     .addFileFixity(
                             "dir1/file3",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "cb6f4f7b3d3eef05d3d0327335071d14c120e065fa43364690fea47d456e146dd334d78d35f73926067d0bf46f122ea026508954b71e8e25c351ff75c993c2b2");
         });
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("3"), updater -> {
@@ -527,11 +528,11 @@ public abstract class OcflITest {
                     .addPath(ITestHelper.sourceObjectPath(objectId, "v3"))
                     .addFileFixity(
                             "file2",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")
                     .addFileFixity(
                             "file4",
-                            DigestAlgorithm.sha512,
+                            DigestAlgorithmRegistry.sha512,
                             "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc");
         });
 
@@ -551,13 +552,13 @@ public abstract class OcflITest {
                                 "file1",
                                 O1_PATH + "/v1/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "96a26e7629b55187f9ba3edc4acc940495d582093b8a88cb1f0303cf3399fe6b1f5283d76dfd561fc401a0cdf878c5aad9f2d6e7e2d9ceee678757bb5d95c39e")),
                         fileDetails(
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6"))));
 
         assertThat(
@@ -570,19 +571,19 @@ public abstract class OcflITest {
                                 "file1",
                                 O1_PATH + "/v2/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc")),
                         fileDetails(
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         fileDetails(
                                 "dir1/file3",
                                 O1_PATH + "/v2/content/dir1/file3",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "cb6f4f7b3d3eef05d3d0327335071d14c120e065fa43364690fea47d456e146dd334d78d35f73926067d0bf46f122ea026508954b71e8e25c351ff75c993c2b2"))));
 
         assertThat(
@@ -595,13 +596,13 @@ public abstract class OcflITest {
                                 "file2",
                                 O1_PATH + "/v1/content/file2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         fileDetails(
                                 "file4",
                                 O1_PATH + "/v2/content/file1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "aff2318b35d3fbc05670b834b9770fd418e4e1b4adc502e6875d598ab3072ca76667121dac04b694c47c71be80f6d259316c7bd0e19d40827cb3f27ee03aa2fc"))));
 
         assertSame(objectDetails.getHeadVersion(), objectDetails.getVersion(VersionNum.fromString("v3")));
@@ -617,7 +618,7 @@ public abstract class OcflITest {
         OcflAsserts.assertThrowsWithMessage(FixityCheckException.class, "Expected md5 digest of", () -> {
             repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("1"), updater -> {
                 updater.addPath(ITestHelper.sourceObjectPath(objectId, "v1"))
-                        .addFileFixity("file1", DigestAlgorithm.md5, "bogus");
+                        .addFileFixity("file1", DigestAlgorithmRegistry.md5, "bogus");
             });
         });
     }
@@ -630,7 +631,9 @@ public abstract class OcflITest {
         var objectId = "o1";
 
         OcflAsserts.assertThrowsWithMessage(
-                OcflInputException.class, "specified digest algorithm is not supported", () -> {
+                OcflInputException.class,
+                "The specified digest algorithm is not mapped to a Java name: DigestAlgorithm{ocflName='bogus', javaStandardName='null'}",
+                () -> {
                     repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("1"), updater -> {
                         updater.addPath(ITestHelper.sourceObjectPath(objectId, "v1"))
                                 .addFileFixity("file1", DigestAlgorithm.fromOcflName("bogus"), "bogus");
@@ -653,7 +656,7 @@ public abstract class OcflITest {
             repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo.setMessage("2"), updater -> {
                 updater.clearVersionState()
                         .addPath(ITestHelper.sourceObjectPath(objectId, "v2"))
-                        .addFileFixity("file2", DigestAlgorithm.md5, "55c1824fcae2b1b51cef5037405fc1ad");
+                        .addFileFixity("file2", DigestAlgorithmRegistry.md5, "55c1824fcae2b1b51cef5037405fc1ad");
             });
         });
     }
@@ -679,14 +682,14 @@ public abstract class OcflITest {
                                 O2_PATH + "/v1/content/dir1/dir2/file2",
                                 "Test file 2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         versionFile(
                                 "dir1/file3",
                                 O2_PATH + "/v3/content/dir1/file3",
                                 "This is a different file 3",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "6e027f3dc89e0bfd97e4c2ec6919a8fb793bdc7b5c513bea618f174beec32a66d2fc0ce19439751e2f01ae49f78c56dcfc7b49c167a751c823d09da8419a4331"))));
 
         ocflObject = repo.getObject(ObjectVersionId.version(objectId, "v1"));
@@ -702,14 +705,14 @@ public abstract class OcflITest {
                                 O2_PATH + "/v1/content/dir1/dir2/file2",
                                 "Test file 2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         versionFile(
                                 "file1",
                                 O2_PATH + "/v1/content/file1",
                                 "Test file 1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "96a26e7629b55187f9ba3edc4acc940495d582093b8a88cb1f0303cf3399fe6b1f5283d76dfd561fc401a0cdf878c5aad9f2d6e7e2d9ceee678757bb5d95c39e"))));
     }
 
@@ -747,7 +750,7 @@ public abstract class OcflITest {
                                 O1_PATH + "/v1/content/f1",
                                 versionInfo(defaultVersionInfo.getUser(), "1"),
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4dff4ea340f0a823f15d3f4f01ab62eae0e5da579ccb851f8db9dfe84c58b2b37b89903a740e1ee172da793a6e79d560e5f7f9bd058a12a280433ed6fa46510a")),
                         fileChange(
                                 FileChangeType.REMOVE,
@@ -763,7 +766,7 @@ public abstract class OcflITest {
                                 O1_PATH + "/v1/content/f1",
                                 versionInfo(defaultVersionInfo.getUser(), "3"),
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4dff4ea340f0a823f15d3f4f01ab62eae0e5da579ccb851f8db9dfe84c58b2b37b89903a740e1ee172da793a6e79d560e5f7f9bd058a12a280433ed6fa46510a"))));
 
         assertThat(
@@ -776,7 +779,7 @@ public abstract class OcflITest {
                                 O1_PATH + "/v1/content/f2",
                                 versionInfo(defaultVersionInfo.getUser(), "1"),
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "40b244112641dd78dd4f93b6c9190dd46e0099194d5a44257b7efad6ef9ff4683da1eda0244448cb343aa688f5d3efd7314dafe580ac0bcbf115aeca9e8dc114")),
                         fileChange(
                                 FileChangeType.UPDATE,
@@ -785,7 +788,7 @@ public abstract class OcflITest {
                                 O1_PATH + "/v3/content/f2",
                                 versionInfo(defaultVersionInfo.getUser(), "3"),
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "7db70149dac5561e411a202629d06832b06b7e8dfef61086ff9e0922459fbe14a69d565cf838fd43681fdb29a698bfe377861b966d12416298997843820bfdb7"))));
 
         assertThat(
@@ -798,7 +801,7 @@ public abstract class OcflITest {
                                 O1_PATH + "/v2/content/f3",
                                 versionInfo(defaultVersionInfo.getUser(), "2"),
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "3bafbf08882a2d10133093a1b8433f50563b93c14acd05b79028eb1d12799027241450980651994501423a66c276ae26c43b739bc65c4e16b10c3af6c202aebb"))));
     }
 
@@ -849,14 +852,14 @@ public abstract class OcflITest {
                                 O2_PATH + "/v1/content/dir1/dir2/file2",
                                 "Test file 2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         versionFile(
                                 "dir1/file3",
                                 O2_PATH + "/v3/content/dir1/file3",
                                 "This is a different file 3",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "6e027f3dc89e0bfd97e4c2ec6919a8fb793bdc7b5c513bea618f174beec32a66d2fc0ce19439751e2f01ae49f78c56dcfc7b49c167a751c823d09da8419a4331"))));
 
         objectVersion = repo.getObject(ObjectVersionId.version(objectId, "v1"));
@@ -872,14 +875,14 @@ public abstract class OcflITest {
                                 O2_PATH + "/v1/content/dir1/dir2/file2",
                                 "Test file 2",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "4cf0ff5673ec65d9900df95502ed92b2605fc602ca20b6901652c7561b302668026095813af6adb0e663bdcdbe1f276d18bf0de254992a78573ad6574e7ae1f6")),
                         versionFile(
                                 "file1",
                                 O2_PATH + "/v1/content/file1",
                                 "Test file 1",
                                 Map.of(
-                                        DigestAlgorithm.sha512,
+                                        DigestAlgorithmRegistry.sha512,
                                         "96a26e7629b55187f9ba3edc4acc940495d582093b8a88cb1f0303cf3399fe6b1f5283d76dfd561fc401a0cdf878c5aad9f2d6e7e2d9ceee678757bb5d95c39e"))));
     }
 
@@ -899,7 +902,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void removeAllOfTheFilesFromAnObject() throws IOException {
+    public void removeAllOfTheFilesFromAnObject() {
         var repoName = "repo4";
         var repo = defaultRepo(repoName);
 
@@ -922,7 +925,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void rejectObjectNotFoundWhenObjectDoesNotExists() throws IOException {
+    public void rejectObjectNotFoundWhenObjectDoesNotExists() {
         var repoName = "repo4";
         var repoDir = ITestHelper.expectedRepoPath(repoName);
         var repo = existingRepo(repoName, repoDir);
@@ -933,7 +936,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void rejectObjectNotFoundWhenObjectExistsButVersionDoesNot() throws IOException {
+    public void rejectObjectNotFoundWhenObjectExistsButVersionDoesNot() {
         var repoName = "repo4";
         var repoDir = ITestHelper.expectedRepoPath(repoName);
         var repo = existingRepo(repoName, repoDir);
@@ -1536,7 +1539,7 @@ public abstract class OcflITest {
             updater.writeFile(
                     new FixityCheckInputStream(
                             inputStream(sourcePath.resolve("file1")),
-                            DigestAlgorithm.md5,
+                            DigestAlgorithmRegistry.md5,
                             "95efdf0764d92207b4698025f2518456"),
                     "file1");
         });
@@ -1557,7 +1560,7 @@ public abstract class OcflITest {
             repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo, updater -> {
                 updater.writeFile(
                         new FixityCheckInputStream(
-                                inputStream(sourcePath.resolve("file1")), DigestAlgorithm.md5, "bogus"),
+                                inputStream(sourcePath.resolve("file1")), DigestAlgorithmRegistry.md5, "bogus"),
                         "file1");
             });
         });
@@ -2408,7 +2411,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void corruptObjectWhenInvalidDigestProvided() throws IOException {
+    public void corruptObjectWhenInvalidDigestProvided() {
         var repoName = "unsafe";
         var repo = defaultRepo(repoName);
 
@@ -2436,7 +2439,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void failCommitWhenMissingFile() throws IOException {
+    public void failCommitWhenMissingFile() {
         var repoName = "unsafe";
         var repo = defaultRepo(repoName);
 
@@ -2474,7 +2477,7 @@ public abstract class OcflITest {
     }
 
     @Test
-    public void failCommitWhenHasExtraFile() throws IOException {
+    public void failCommitWhenHasExtraFile() {
         var repoName = "unsafe";
         var repo = defaultRepo(repoName);
 
@@ -2766,6 +2769,67 @@ public abstract class OcflITest {
         assertEquals(OcflVersion.OCFL_1_1, repo.describeObject(objId).getObjectOcflVersion());
         var result = repo.validateObject(objId, true);
         assertEquals(0, result.getErrors().size());
+    }
+
+    @Test
+    public void sizeBasedFixityShouldValidateWhenCorrect() throws IOException {
+        var repoName = "size-fixity-valid";
+        var repo = defaultRepo(repoName);
+
+        var objectId = "obj1";
+
+        var file = writeFile("asdf".repeat(99999));
+        var fileSize = Files.size(file);
+
+        repo.updateObject(ObjectVersionId.head(objectId), null, updater -> {
+            updater.writeFile(ITestHelper.streamString("file1"), "file1");
+            updater.addFileFixity("file1", DigestAlgorithmRegistry.size, "5");
+            updater.addPath(file, "file2");
+            updater.addFileFixity("file2", DigestAlgorithmRegistry.size, String.valueOf(fileSize));
+        });
+
+        assertEquals(
+                "5",
+                repo.getObject(ObjectVersionId.head(objectId))
+                        .getFile("file1")
+                        .getFixity()
+                        .get(DigestAlgorithmRegistry.size));
+        assertEquals(
+                String.valueOf(fileSize),
+                repo.getObject(ObjectVersionId.head(objectId))
+                        .getFile("file2")
+                        .getFixity()
+                        .get(DigestAlgorithmRegistry.size));
+
+        var results = repo.validateObject(objectId, false);
+
+        assertEquals(0, results.getErrors().size(), () -> results.getErrors().toString());
+    }
+
+    @Test
+    public void sizeBasedFixityShouldFailValidationWhenInvalid() {
+        var repoName = "size-fixity";
+        var repo = defaultRepo(repoName);
+
+        var objectId = "obj1";
+
+        var file = writeFile("asdf".repeat(99999));
+
+        OcflAsserts.assertThrowsWithMessage(
+                FixityCheckException.class, "Expected size digest of file1 to be 100, but was 5.", () -> {
+                    repo.updateObject(ObjectVersionId.head(objectId), null, updater -> {
+                        updater.writeFile(ITestHelper.streamString("file1"), "file1");
+                        updater.addFileFixity("file1", DigestAlgorithmRegistry.size, "100");
+                    });
+                });
+
+        OcflAsserts.assertThrowsWithMessage(
+                FixityCheckException.class, "Expected size digest of file2 to be 100, but was 399996.", () -> {
+                    repo.updateObject(ObjectVersionId.head(objectId), null, updater -> {
+                        updater.addPath(file, "file2");
+                        updater.addFileFixity("file2", DigestAlgorithmRegistry.size, "100");
+                    });
+                });
     }
 
     private Path writeFile(String content) {